Crate cargo_wix

cargo-wix Binary and Subcommand

The goal of the cargo-wix project and the cargo wix subcommand is to make it easy to create a Windows installer (msi) for any Rust project. The project is primarily implemented as a cargo subcommand, but the core functionality is provided in a library (crate). See the module-level comments for the library for more information about usage and organization of the wix crate. The remainder of this documentation focuses on the usage and features of the cargo wix subcommand.

Quick Start

Examples

Features

Flags and Options

Quick Start

Ensure the WiX Toolset is installed and a WIX system environment variable has been created. The installer for the WiX Toolset should create the WIX system environment variable automatically. Then, start, or restart, a command prompt (cmd.exe) and execute the following commands:

C:\>cargo install cargo-wix
C:\>cd Path\To\Project
C:\Path\To\Project\>cargo wix init
C:\Path\To\Project\>cargo wix

The Windows installer (msi) will be in the C:\Path\To\Project\target\wix folder. The cargo wix init command will create a wix folder alongside the src folder and the project's manifest (Cargo.toml). The wix folder will contain the WiX Source file (main.wxs) that was automatically generated for the project based the contents of its manifest. The WiX Source file can be customized using a text editor, and once the file exists, the cargo wix init command does not need to be used again.

The cargo wix command uses the wix\main.wxs file generated from the previous cargo wix init command as input for the WiX Toolset's "compiler" and "linker", i.e. candle.exe and light.exe, respectively, to create the Windows installer (msi). A variety of artifact files will be created in the target\wix folder. These can be ignored and/or deleted.

The created installer will install the executable file in a bin folder within the destination selected by the user during installation. It will add a license file to the same folder as the bin folder, and it will add the bin folder to the PATH system environment variable so that the executable can be called from anywhere with a command prompt. Most of these behaviors can be adjusted during the installation process. The default installation destination is C:\Program Files\<project name>, where <project name> is replaced with the name of the project.

Examples

All of the following examples use the native Command Prompt (cmd.exe) for the Windows OS; however, the Developer Prompt installed with the VC Build Tools is recommended. The git bash terminal can also be used.

Begin each example by staring the appropriate prompt and navigating to the root folder of the Rust project. The cargo wix subcommand and binary assumes the current working directory (cwd) is the project, a.k.a. package, root folder, i.e. the same folder as the package's manifest (Cargo.toml).

Let's create a basic project with Cargo and then an installer.

C:\Path\to\Project> cargo init --bin --vcs none --name example

This will create a simple binary package named "example" without any verison control. While the cargo wix subcommand and binary does work with libraries (crates) in addition to binaries (applications), generally an installer is needed for a binary (application) not a library. Rust libraries are typically distributed via crates.io and do not need an installer. The package's manifest should look like the following, but with your name and email address for the authors field:

[package]
name = "example"
version = "0.1.0"
authors = ["First Last <first.last@example.com>"]

[dependencies]

The next step is to create the WiX Source (wxs) file, which is needed to create the installer for the project.

C:\Path\to\Project> cargo wix init
 WARN: A description was not specified at the command line or in the package's manifest (Cargo.toml). The description can be added manually to the generated WiX Source (wxs) file using a text editor.
 WARN: An EULA was not specified at the command line, a RTF license file was not specified in the package manifest's (Cargo.toml) 'license-file' field, or the license ID from the pacakge manifest's 'license' field is not recognized. The license agreement dialog will be excluded from the installer. An EULA can be added manually to the generated WiX Source (wxs) file using a text editor.
 WARN: A help URL could not be found and it will be excluded from the installer. A help URL can be added manually to the generated WiX Source (wxs) file using a text editor.
 WARN: A license file could not be found and it will be excluded from the installer. A license file can be added manually to the generated WiX Source (wxs) file using a text editor.

The warnings can be ignored for the time being, and they will be addressed in subsequent examples. The above command will create a wix folder with the main.wxs file:

C:\Path\to\Project> dir /B
Cargo.toml
src
wix
C:\Path\to\Project> dir wix /B
main.wxs

The wix folder and the main.wxs file now exist, and an installer can be created:

C:\Path\to\Project> cargo wix

This may take a moment to complete as the cargo wix subcommand will build the application with the Release target and then build the installer. The installer will be located in the target\wix folder.

C:\Path\to\Project> dir target\wix /B
example-0.1.0-x86_64.msi
example-0.1.0-x86_64.wixpdb
main.wixobj

Great! An installer (msi) exists for the application. The example-0.1.0-x86_54.wixpdb and main.wixobj files are artifacts of the installer build process and can be ignored and/or deleted.

The installer that is created with the above steps and commands will install the example.exe file to: C:\Program Files\example\bin\example.exe. It will also add the C:\Program Files\example\bin path to the PATH system environment variable, unless this feature is disabled during the installation process from the Custom Setup dialog. The default welcome screen, banner, and icons from the WiX Toolset's default UI component will be used. A license dialog will not appear for this installer.

The installer that is created from the previous example is relatively simple, but so is the application and package. It would be nice to have a license dialog appear in the installer that allows the end-user to review and acknowledge the End User License Agreement (EULA). It would also be nice to have this same license appear in the installation folder for the application so end-users can review it after installation.

The license agreement dialog for a WiX Toolset created installer must be in the Rich Text Format (.rtf). The majority of Rust binaries and libraries are developed and distributed with an open source license. We will do the same for the example binary here and use the GPL-3.0 license. Creating a RTF version of the GPL-3.0 requires a third-party tool, such as WordPad, which is freely available with any Windows distribution, Microsoft Office, LibreOffice, or any number of freely available text editors, word processors, or document conversion tools. Luckily, the cargo wix subcommand and binary has a RTF version of the GPL-3.0 license available and eliminates the need to install, pay, and/or learn another tool to create the EULA.

First, open the package's manifest (Cargo.toml) in a text editor, like Microsoft Notepad, and add the license field to the package section with the GPL-3.0 value:

[package]
name = "example"
version = "0.1.0"
authors = ["First Last <first.last@example.com>"]
license = "GPL-3.0"

[dependencies]

Save the package's manifest and exit the text editor. Now, we can create a wix\main.wxs file that will include the license dialog with a GPL-3.0 EULA.

C:\Path\to\Project> cargo wix init --force
 WARN: A description was not specified at the command line or in the package's manifest (Cargo.toml). The description can be added manually to the generated WiX Source (wxs) file using a text editor.
 WARN: A help URL could not be found and it will be excluded from the installer. A help URL can be added manually to the generated WiX Source (wxs) file using a text editor.

C:\Path\to\Project> dir wix /B
License.rtf
main.wxs
C:\Path\to\Project> cargo wix

The --force flag is needed so that the existing wix\main.wxs is overwritten with the new EULA-enabled wix\main.wxs file. The majority of the warnings have also been addressed. Notice there is now a License.rtf file in the wix folder. This will be used as the EULA in the license agreement dialog for the installer and be included in the installation folder with the binary.

A side note, while version control has been disabled for these examples, it is best practice to include installer-related files in version control; thus, the wix\main.wxs and wix\License.rtf should be added to version control.

Let's fix the remaining warnings. Both of the remaining warnings can be resolved in multiple ways. The first is to use the options available for the cargo wix init subcommmand following the previous example project:

C:\Path\to\Project> cargo wix init --force -d "This is a description" -u http://www.example.com

C:\Path\to\Project> dir wix /B
License.rtf
main.wxs
C:\Path\to\Project> cargo wix

The warnings for the description and help URL have disappeared. The -d,--description option for the cargo wix init command adds a description for the installer. Similarly, the -u,--url option adds a help URL. The --force flag is still needed to overwrite the previous wix\main.wxs file.

Another possibility is to add the description and homepage fields to the package's manifest (Cargo.toml), and then initialize and create the installer. The cargo-wix binary and subcommand will use these fields, among others, to automatically include the values into the installer. The -d,--description and -u,--url options can still be used to override the values from the package's manifest. This can be useful if the contents of the installer might need to be different or more detailed than the package.

Following from the previous example, open the package's manifest (Cargo.toml) in a text editor, like Microsoft Notepad, and add the description and homepage fields to the package section:

[package]
name = "example"
version = "0.1.0"
authors = ["First Last <first.last@example.com>"]
license = "GPL-3.0"
description = "This is a description"
homepage = "http://www.example.com"

[dependencies]

Save the package's manifest and exit the text editor. Now, we can create a wix\main.wxs file witout any warnings and uses the description and homepage from the package's manifest:

C:\Path\to\Project> cargo wix init --force
C:\Path\to\Project> dir wix /B
License.rtf
main.wxs
C:\Path\to\Project> cargo wix

The documentation and repository fields can be used instead of the homepage field for the help URL, too.

Features

The cargo-wix binary, and related cargo wix subcommand, use the WiX Toolset and the SignTool application available in the Windows 10 SDK. These are obviously Windows-only applications, so while the crate and binary can be built on any platform supported by the Rust programming language and Cargo, the cargo wix subcommand is only really useful in a Windows environment.

The WiX Toolset provides a "compiler" (candle.exe) and "linker" (light.exe). These can be found in the bin directory of the installation location for the WiX Toolset. The value of the WIX system environment variable that is created during installation of the WiX Toolset is a path to the installation folder that contains the bin folder. The WIX system environment variable is used by the cargo wix subcommand with the std::process::Command module to create installers.

The Windows SDK provides a signer (signtool) application for signing installers. The application is installed in the bin folder of the Windows SDK installation. The location of the bin folder varies depending on the version. It is recommended to use the Developer Prompt to ensure the signtool application is available. Signing an installer is optional.

The WiX Toolset requires a WiX Source (wxs) file, which is an XML file. A template is provided with this binary that attempts to meet the majority of use cases for developers and avoid requiring extensive knowledge of the WiX Toolset and Windows installer technologies. Modification of the template is encouraged, but please consult the WiX Toolset's extensive documentation and tutorials for information about writing (authoring), customizing, and using wxs files. This documentation here is only for this binary and subcommand.

The WXS template is embedded in the binary installation of the subcommand and it can be printed to STDOUT using the cargo wix print wxs command from the command prompt (cmd.exe). Note, each time the cargo wix print wxs command is invoked, new GUIDs are generated for fields that require them. Thus, a developer does not need to worry about generating GUIDs and can begin using the template immediately with this subcommand or the WiX Toolset's compiler (candle.exe) and linker (light.exe) applications.

In addition to the WXS template, there are several license templates which are used to generate an End User License Agreement (EULA) during the cargo wix init command. Depending on the license ID(s) in the license field for a package's manifest (Cargo.toml), a license file in the Rich Text Format (RTF) is generated from a template and placed in the wix folder. This RTF file is then displayed in the license agreement dialog of the installer. See the help information on the carge wix print subcommand:

C:\Path\to\Project> cargo wix print --help

for information about supported licenses. If the license field is not used, or the license ID is not supported, then the EULA is not automatically created during initialization and it will have to be created manually with a text editor or some other authoring tool.

The cargo wix init subcommand uses a combination of the license and license-file fields of the project's manifest (Cargo.toml) to determine if a sidecar license file should be included in the installation folder alongside the bin folder. The license field appears to be the more commonly used field to describe the licensing for a Rust project and package, while the license-file field is used to specify a custom, or properitary, license. The top three most common licenses for Rust projects are supported from the license field, i.e. MIT, Apache-2.0, and GPLv3. If any of these three supported open source licenses are used for the license field, then a License.rtf file is generated from an embedded template and placed in the wix folder as part of the cargo wix init subcommand. This generated RTF file will be used as a sidecar file and for the End User License Agreement (EULA) that is displayed in the license agreement dialog of the installer. If the license-file field is used and it contains a path to a file with the .rtf extension, then this file will be used as a sidecar file and for the EULA. If neither of these fields exist or contain valid values, then no sidecar file is included in the installation and no license agreement dialog appears during installation. This default behavior can be overridden with the -l,--license and -e,--eula options for the cargo wix init subcommand.

Flags and Options

Generally, any value that is obtained from the package's manifest (Cargo.toml) can be overridden at the command line with an appropriate option. For example, the manufacturer, which is displayed as the "Publisher" in the Add/Remove Programs (ARP) control panel is obtained from the first author listed in the authors field of a project's manifest, but it can be overridden using the -m,--manufacturer option with the cargo wix init subcommand.

Use the -h,--help flag with each subcommand to get a full list of options and flags that are available. The short flag, -h, will print a condensed version of each flag and option for the subcommand, while the long flag, --help, will print a detailed help for each flag and option. The rest of this section is a list of all flags and options implemented for all subcommands.

-b,--banner

Available for the init (cargo wix init) and print (cargo wix print) subcommands.

Sets the path to a bitmap (.bmp) image file that will be displayed across the top of each dialog in the installer. The banner image dimensions should be 493 x 58 pixels.

-b,--bin-path

Available for the default (cargo wix) and sign (cargo wix sign) subcommands.

The -b,--bin-path option can be used to specify a path (relative or absolute) to the WiX Toolset bin folder. The -b,--bin-path option is useful if a different version of the WiX Toolset needs to be used to create the installer. The descending order of precedence is: (1) -b,--bin-path option then (2) WIX system environment variable. An error will be displayed if the compiler and/or linker cannot be found.

This option is also avaliable for the cargo wix sign subcommand and can be used to specify a path to the Windows SDK bin folder. This can be used to override default signtool application found using the std::process::Command::status method.

-B,--binary

Available for the init (cargo wix init) and print (cargo wix print) subcommands.

The binary name, which is the name field for the [[bin]] section, is used for the executable file name, i.e. "name.exe". This can also be overridden using the -B,--binary option for the cargo wix init subcommand. This option takes a relative or absolute path to an executable file and uses the file name as the binary name.

-c,--culture

Available for the default (cargo wix) subcommand.

Sets the culture for localization. Use with the [-l,--locale] option. See the WixUI localization documentation for more information about acceptable culture codes. The codes are case insensitive.

-d,--description

Available for the init (cargo wix init), print (cargo wix print), and sign (cargo wix sign) subcommands.

The package description is used in multiple places for the installer, including the text that appears in the blue UAC dialog when using a signed installer. This can be overridden using the -d,--description option with the cargo wix init or cargo wix sign subcommands, respectively.

-D,--dialog

Available for the init (cargo wix init) and print (cargo wix print) subcommands. Sets the path to a bitmap (.bmp) image file that will be displayed to the left on the first dialog of the installer. The dialog image dimensions should be 493 x 312 pixels. The first dialog is known as the "Welcome" dialog.

-e,--eula

Available for the init (cargo wix init) and print (cargo wix print) subcommands.

Specifies a Rich Text Format (RTF) file to use as the End User License Agreement (EULA) for the license agreement dialog of the installer. The default is to disable the license agreement dialog unless one of the supported licenses (GPL-3.0, Apache-2.0, or MIT) is generated based on the value of the license field in the package's manifest (Cargo.toml). An EULA can be enabled later by directly modifying the WiX Source (wxs) file with a text editor.

--force

Available for the init (cargo wix init) subcommand.

Forces overriding of generated files from the init subcommand. Use with caution! This cannot be undone.

-h,--help

Available for all subcommands.

The short flag, -h, will print a condensed version of the help text, while the long flag, --help, will print a more detailed version of the help text.

-u,--homepage

Available for the sign (cargo wix sign) subcommand.

This will be display in the ACL dialog.

-i,--install-version

Available for the default (cargo wix) subcommand.

Overrides the version from the package's manifest (Cargo.toml), which is used for the installer name and appears in the Add/Remove Programs (ARP) control panel.

-l,--license

Available for the init (cargo wix init) and print (cargo wix print) subcommands.

Overrides the license-file field or any license generated from the license field of the package's manifest (Cargo.toml). If an appropriate license file does not exist, cannot be found, or is not specified, then no license file is included in the installer or in the installation folder along side the bin folder. A file containing the license, such as a TXT, PDF, or RTF file, can be added later by directly editing the generated WiX Source file (wxs) in a text editor.

-l,--locale

Available for the default (cargo wix) subcommand.

Sets the path to a WiX localization file (wxl) which contains localized strings. Use in conjunction with the [-c,--culture] option.

-m,--manufacturer

Available for the init (cargo wix init) and print (cargo wix print) subcommands.

Overrides the first author in the authors field of the package's manifest (Cargo.toml) as the manufacturer within the installer. The manufacturer can be changed after initialization by directly modifying the WiX Source file (wxs) with a text editor.

-n,--name

Available for the default (cargo wix) subcommand.

Overrides the name field in the package's manifest (Cargo.toml), which is used in the file name of the installer (msi). This does not change the name of the executable within the installer.

--no-build

Available for the default (cargo wix) subcommand.

This skips building the Rust package using Cargo for the Release target.

--nocapture

Available for the default (cargo wix) and sign (cargo wix sign) subcommands.

Displays all output from the builder (Cargo), compiler (candle.exe), linker (light.exe), and signer (signtool.exe) applications.

-o,--output

Available for the default (cargo wix), init (cargo wix init) and print (cargo wix print) subcommands.

Sets the destination for init subcommand files, such as the WiX Source file (wxs), an alternative to stdout for print subcommand, and the created installer for the default subcommand.

-O,--owner

Available for the init (cargo wix init) and print (cargo wix print) subcommands.

Sets the copyright owner of the license during initialization or printing. The default is to use the first author from the package's manifest (Cargo.toml). This is only used when generating a license based on the value of the license field in the package's manifest.

-p,--product-icon

Available for the init (cargo wix init) and print (cargo wix print) subcommands.

Sets the path to an image file that will be display as an icon in the Add/Remove Programs (ARP) control panel for the installed application.

-P,--product-name

Available for the init (cargo wix init), print (cargo wix print), and sign (cargo wix sign) subcommands.

Overrides the name field of the package's manifest (Cargo.toml) as the product name within the installer. The product name can be changed after initialization or printing by directly modifying the WiX Source file (wxs) with a text editor.

-t,--timestamp

Available for the sign (cargo wix sign) subcommand.

An alias or URL to a timestamp server when signing an installer with a certificate. Valid aliases are: Comodo and Versign, which are case insenstive.

-u,--url

Available for the init (cargo wix init) and print (cargo wix print) subcommands.

Adds a URL to the installer that will be displayed in the Add/Remove Programs (ARP) control panel for the application. The default is to disable it unless a URL is specified for either the homepage, documentation, or repository fields in the package's manifest (Cargo.toml). The help URL can be enabled after initialization by directly modifying the WiX Source (wxs) file with a text editor.

-V,--version

Available for all subcommands.

Prints the cargo-wix binary and subcommand version.

-v,--verbose

Available for all subcommands.

Increases the level of logging statements based on occurance count of the flag. The more -v,--verbose flags used, the more logging statements that will be printed during execution of a subcommand. When combined with the --nocapture flag, this is useful for debugging and testing.

-y,--year

Available for the init (cargo wix init) and print (cargo wix print) subcommands.

Sets the copyright year for the license during initialization. The default is to use the current year. This is only used if a license is generated from one of the supported licenses based on the value of the license field in the package's manifest (Cargo.toml).