Enum cargo_wix::Error

pub enum Error {
    Command(&'static str, i32),
    Generic(String),
    Io(Error),
    Manifest(String),
    Mustache(Error),
    Toml(Error),
}

The error type for cargo-wix-related operations and associated traits.

Errors mostly originate from the dependencies, but custom instances of Error can be created with the Generic variant and a message.

Variants
Command(&'static str, i32)

A command operation failed.

Generic(String)

A generic or custom error occurred. The message should contain the detailed information.

Io(Error)

An I/O operation failed.

Manifest(String)

A needed field within the Cargo.toml manifest could not be found.

Mustache(Error)

An error occurred with rendering the template using the mustache renderer.

Toml(Error)

Parsing of the Cargo.toml manifest failed.

Methods

impl Error

pub fn code(&self) -> i32

Gets an error code related to the error.

This is useful as a return, or exit, code for a command line application, where a non-zero integer indicates a failure in the application. it can also be used for quickly and easily testing equality between two errors.