Enum cargo_wix::Platform

pub enum Platform {
    X86,
    X64,
}

The different values for the Platform attribute of the Package element.

Variants
X86

The x86 WiX Toolset value.

X64

The x64 WiX Toolset value.

Methods

impl Platform

pub fn arch(&self) -> &'static str

Gets the name of the platform as an architecture string as used in Rust toolchains.

This is different from the string used in WiX Source (wxs) files. This is the string commonly used for the target_arch conditional compilation attribute. To get the string recognized in wxs files, use format!("{}", Platform::X86).

Example

extern crate cargo_wix;
 
use cargo_wix::Platform;

fn main() {
    assert_eq!(Platform::X86.arch(), "i686");
    assert_eq!(Platform::X64.arch(), "x86_64");
}