Enum cargo_wix::Template

pub enum Template {
    Apache2,
    Gpl3,
    Mit,
    Wxs,
}

The different templates that can be printed using the --print-template option.

Variants
Apache2Gpl3MitWxs
Methods

impl Template

pub fn id(&self) -> &str

Gets the ID for the template.

In the case of a license template, the ID is the SPDX ID which is also used for the license field in the package's manifest (Cargo.toml). This is also the same value used with the --print-template option.

pub fn possible_values() -> Vec<String>

Gets the possible string representations of each variant.

pub fn to_str(&self) -> &str

Gets the embedded contents of the template as a string.